refactor(pages): move pages into separate folders with their styles
