+++ src/pages/AlbumPhotosPage/AlbumPhotosPage.tsx
@@ -1,7 +1,7 @@
 import type { FC } from "react";
 import { useParams } from "react-router-dom";
-import { useTheme } from "../shared/lib/theme/useTheme";
-import { mockPhotos } from "../shared/mocks/photos";
+import { useTheme } from "../../shared/lib/theme/useTheme";
+import { mockPhotos } from "../../shared/mocks/photos";
 import styles from "./AlbumPhotosPage.module.css";
 import clsx from "clsx";
 
@@ -9,7 +9,9 @@ export const AlbumPhotosPage: FC = () => {
   const { albumId } = useParams<{ albumId: string }>();
   const { theme } = useTheme();
 
-  const photos = mockPhotos.filter((photo) => photo.albumId === Number(albumId));
+  const photos = mockPhotos.filter(
+    (photo) => photo.albumId === Number(albumId),
+  );
 
   if (photos.length === 0) {
     return <div>No photos found for this album.</div>;
