+++ src/pages/PostDetailsPage/PostDetailsPage.tsx
@@ -1,12 +1,12 @@
 import { useMemo, type FC } from "react";
 import { Link, useParams } from "react-router-dom";
-import { mockPosts } from "../shared/mocks/posts";
-import { mockComments } from "../shared/mocks/comments";
-import { CommentList } from "../widgets/CommentList/ui/CommentList";
-import { useTheme } from "../shared/lib/theme/useTheme";
+import { mockPosts } from "../../shared/mocks/posts";
+import { mockComments } from "../../shared/mocks/comments";
+import { CommentList } from "../../widgets/CommentList/ui/CommentList";
+import { useTheme } from "../../shared/lib/theme/useTheme";
 import styles from "./PostDetailsPage.module.css";
-import { mockUsers } from "../shared/mocks/users";
-import type { User } from "../entities/post/model/types";
+import { mockUsers } from "../../shared/mocks/users";
+import type { User } from "../../entities/post/model/types";
 
 export const PostDetailsPage: FC = () => {
   const { id } = useParams<{ id: string }>();
